va_arg works. idk why results are different

+++ src/s21_string.h
@@ -8,17 +8,17 @@
 #include <math.h>
 
 typedef union {
-    float full;
-    unsigned bits;
+	double full;
+	unsigned long long bits;
 } f_representation;
 
 int s21_sprintf(char *str, const char *format, ...);
 void specifier(const char *format, int i, char *str, int *j, va_list *params);
-char *doxtoa(int d, const int radix, const bool uppercase);
-int doxlen(int d, const int radix);
+char *doxtoa(long long d, const int radix, const bool uppercase);
+int doxlen(long long d, const int radix);
 int etoa(char* f_str);
-char *ftoa(float f);
-int extract_exp(unsigned bits);
+char *ftoa(double f);
+int extract_exp(unsigned long long bits);
 char* add_rank(char *str, int num, char value);
 char *addnulles(char *str, int i, int j);
 char* stradd(char *l_str, char *r_str, bool fraction);
